+++ src/WillHero/Coin.java
@@ -4,7 +4,7 @@ import javafx.scene.image.ImageView;
 
 public class Coin extends GameComponents {
 
-    public Coin(float x, float y) {
+    public Coin(double x,double y) {
         super(x, y);
     }
