+++ src/WillHero/fallingPlatform.java
@@ -5,7 +5,7 @@ import javafx.scene.image.ImageView;
 public class fallingPlatform extends Platform{
     private int falling_Time;
 
-    public fallingPlatform(float x, float y, int size,int fallingtime) {
+    public fallingPlatform(double x,double y, int size,int fallingtime) {
         super(x, y, size);
         this.falling_Time = fallingtime;
     }
